requirements.txt
streamlit
Pillow
requests
sseclient-py